// src/app/modules/personal-details/models/personal-details.ts
export interface Profile {
  id?: number;
  name?: string;
  image?: string;
  username?: string;
  email?: string;
  mobile?: string;
  country?: string;
  system_account_code?: string;
  status?: number;
  is_main?: number;
  two_step_status?: number;
  telegram_username?: string;
  telegram_id?: string;
  telegram_code?: string;
  created_at?: string;
  updated_at?: string;
  image_path?: string;
  has_image?: boolean;
  phone?: string;
}